requirements.txt
asteroid_filterbanks
glob2
h5py
intervaltree
matplotlib
numpy
pandas
prettytable
pyannote.pipeline
pyannote.audio
pyannote.core
pyannote.metrics
pytest
PyYAML
scikit_learn
scipy
SoundFile
tabulate
torch
torchaudio
tqdm
